requirements.txt
streamlit
pytesseract
pandas
openpyxl
python-dotenv
Pillow
imaplib2
email-validator
